requirements.txt
fire~=0.4.0
PyYAML~=5.4.1
google-api-python-client~=2.6.0
kubernetes==18.20.0
pytest==6.2.4
psycopg2-binary==2.9.1
